Added CHANGELOG, LICENSE, mint_hashcash.py and verify_hashcash.py
Chose to use a date timestamp for the version numbers, rather than
"1.0", "2.0", etc. This way, it is clear when the latest release is
without checking the webpage for the date.

Also, chose to use the GPLv3 for the license, because Mutt is licensed
under the GPLv2+ and it didn't make sense to give the project a license
that allows proprietary redistribution, when the very project it relies
on does not.

Lastly, the minting and verification Python scripts are inital adds.