Removed old license
When this script was initially created, and I blogged about it, it as
well as the minting script were licensed under the public domain.
However, after creating this project on Github, I've changed the
license to the GPLv3, but forgot to remove the license comment from the
script. Thus, this commit.