Updated logic to close issues 1 & 4
Updated logic to support the tuple of email addresses for verification. By
doing so, not only are tokens only checked for those that have been minted
for the addresses found en EMAILADDR, but any tokens not addressed to said
recipients aren't checked. Also, got rid of the logic displaying that no
tokens were found- figured it was unneeded noise.